+++ src/main/java/com/codetaylor/mc/dropt/modules/dropt/rule/data/RuleMatchSpawnDistance.java
@@ -1,0 +1,11 @@
+package com.codetaylor.mc.dropt.modules.dropt.rule.data;
+
+import com.codetaylor.mc.dropt.api.reference.EnumListType;
+
+public class RuleMatchSpawnDistance {
+
+  public EnumListType type = EnumListType.WHITELIST;
+  public int min = 0;
+  public int max = Integer.MAX_VALUE;
+
+}