+++ docs/changelog.md
@@ -1,5 +1,8 @@
 1.16.0
 
+  * Added:
+    * `matchQuantity` for matching a drop's quantity to an existing drop's quantity [(#55)](https://github.com/codetaylor/dropt/issues/55) [(Examples)](examples/match_quantity.md) [(DOC)](json/syntax/#iruledropmatchquantity)
+
   * Fixed:
     * Edge case that caused undesired behavior if the quantity of item to drop exceeds the item's max stack size